refactor: remove Safe360 page and related components, update links in Navbar and Footer
This commit removes the Safe360 page, its associated components, and styles from the project. The Navbar and Footer have been updated to reflect these changes, with links adjusted to point to the appropriate sections. The AboutFamCare component's link has also been modified for consistency.
