+++ src/components/Safe360Architecture.module.css
@@ -1,0 +1,150 @@
+.section {
+  padding: 108px 0;
+  background:
+    radial-gradient(circle at 86% 12%, rgba(126, 216, 218, 0.14), transparent 28%),
+    #fff;
+  scroll-margin-top: 72px;
+}
+
+.heading {
+  display: grid;
+  grid-template-columns: minmax(0, 1fr) 390px;
+  gap: 60px;
+  align-items: end;
+  margin-bottom: 42px;
+}
+
+.heading h2 {
+  max-width: 760px;
+  font-size: clamp(2.25rem, 3.8vw, 3.45rem);
+  font-weight: 800;
+  line-height: 1.08;
+  letter-spacing: -0.047em;
+}
+
+.heading > p {
+  padding-bottom: 9px;
+  color: var(--ink-60);
+  font-size: 0.96rem;
+  line-height: 1.72;
+}
+
+.grid {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  border: 1px solid rgba(0, 127, 130, 0.16);
+  border-radius: 28px 28px 0 0;
+  background: var(--brand-ghost);
+  overflow: hidden;
+}
+
+.card {
+  min-height: 290px;
+  padding: 28px;
+  background: rgba(255, 255, 255, 0.9);
+  transition: background 0.2s ease;
+}
+
+.card:hover { background: var(--brand-ghost); }
+.card:not(:nth-child(3n)) { border-right: 1px solid rgba(0, 127, 130, 0.12); }
+.card:nth-child(-n + 3) { border-bottom: 1px solid rgba(0, 127, 130, 0.12); }
+
+.cardTop {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.cardTop > span {
+  color: rgba(0, 90, 93, 0.42);
+  font-size: 0.62rem;
+  font-weight: 800;
+  letter-spacing: 0.13em;
+  text-transform: uppercase;
+}
+
+.icon {
+  display: grid;
+  place-items: center;
+  width: 68px;
+  height: 68px;
+  margin: -9px 0 -2px -9px;
+}
+
+.iconArt {
+  width: 100%;
+  height: 100%;
+  object-fit: contain;
+  filter: drop-shadow(0 5px 10px rgba(0, 90, 93, 0.08));
+}
+
+.card h3 {
+  margin-top: 23px;
+  font-size: 1.12rem;
+  font-weight: 800;
+  letter-spacing: -0.025em;
+  text-transform: capitalize;
+}
+
+.card ul {
+  display: grid;
+  gap: 10px;
+  margin-top: 18px;
+  list-style: none;
+}
+
+.card li {
+  display: flex;
+  align-items: center;
+  gap: 9px;
+  color: var(--ink-60);
+  font-size: 0.84rem;
+}
+
+.card li svg { flex: 0 0 auto; color: var(--brand); }
+
+.status {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 10px;
+  min-height: 50px;
+  padding: 10px 20px;
+  border: 1px solid rgba(0, 127, 130, 0.16);
+  border-top: none;
+  border-radius: 0 0 18px 18px;
+  color: var(--brand-dark);
+  background: var(--brand-pale);
+  font-size: 0.74rem;
+  font-weight: 700;
+  text-align: center;
+}
+
+.status span {
+  width: 7px;
+  height: 7px;
+  flex: 0 0 auto;
+  border-radius: 50%;
+  background: var(--brand-mid);
+  box-shadow: 0 0 0 5px rgba(0, 150, 152, 0.1);
+}
+
+@media (max-width: 900px) {
+  .heading { grid-template-columns: 1fr; gap: 10px; }
+  .heading > p { max-width: 650px; padding: 0; }
+  .grid { grid-template-columns: repeat(2, 1fr); }
+  .card:not(:nth-child(3n)) { border-right: none; }
+  .card:nth-child(-n + 3) { border-bottom: none; }
+  .card:nth-child(odd) { border-right: 1px solid rgba(0, 127, 130, 0.12); }
+  .card:nth-child(-n + 4) { border-bottom: 1px solid rgba(0, 127, 130, 0.12); }
+}
+
+@media (max-width: 620px) {
+  .section { padding: 76px 0; }
+  .heading { margin-bottom: 30px; }
+  .grid { grid-template-columns: 1fr; border-radius: 22px 22px 0 0; }
+  .card { min-height: 0; padding: 25px 24px 28px; }
+  .card:nth-child(odd) { border-right: none; }
+  .card:nth-child(-n + 5) { border-bottom: 1px solid rgba(0, 127, 130, 0.12); }
+  .status { line-height: 1.45; }
+}
